Unify `trimToNull` extension function with synctools version
