+++ core/src/main/kotlin/at/bitfire/davdroid/db/Collection.kt
@@ -28,8 +28,8 @@ import at.bitfire.dav4jvm.property.webdav.CurrentUserPrivilegeSet
 import at.bitfire.dav4jvm.property.webdav.DisplayName
 import at.bitfire.dav4jvm.property.webdav.ResourceType
 import at.bitfire.davdroid.util.DavUtils.lastSegment
-import at.bitfire.davdroid.util.trimToNull
 import at.bitfire.ical4android.ICalendar
+import at.bitfire.synctools.util.trimToNull
 import okhttp3.HttpUrl
 import okhttp3.HttpUrl.Companion.toHttpUrlOrNull
