+++ core/src/main/kotlin/at/bitfire/davdroid/db/Principal.kt
@@ -13,7 +13,7 @@ import at.bitfire.dav4jvm.okhttp.UrlUtils
 import at.bitfire.dav4jvm.property.webdav.DisplayName
 import at.bitfire.dav4jvm.property.webdav.ResourceType
 import at.bitfire.dav4jvm.property.webdav.WebDAV
-import at.bitfire.davdroid.util.trimToNull
+import at.bitfire.synctools.util.trimToNull
 import okhttp3.HttpUrl
 
 /**
