+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/calendar/builder/DescriptionBuilder.kt
@@ -6,7 +6,7 @@ package at.bitfire.synctools.mapping.calendar.builder
 
 import android.content.Entity
 import android.provider.CalendarContract.Events
-import at.bitfire.synctools.util.Utils.trimToNull
+import at.bitfire.synctools.util.trimToNull
 import net.fortuna.ical4j.model.component.VEvent
 
 class DescriptionBuilder: AndroidEntityBuilder {
