+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/contacts/ContactReader.kt
@@ -4,7 +4,7 @@
 
 package at.bitfire.synctools.mapping.contacts
 
-import at.bitfire.synctools.util.Utils.trimToNull
+import at.bitfire.synctools.util.trimToNull
 import at.bitfire.synctools.vcard.property.CustomScribes.registerCustomScribes
 import at.bitfire.synctools.vcard.property.CustomType
 import at.bitfire.synctools.vcard.property.XAbDate
