+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/contacts/builder/RelationBuilder.kt
@@ -9,7 +9,7 @@ import android.provider.ContactsContract.CommonDataKinds.Relation
 import at.bitfire.synctools.mapping.contacts.Contact
 import at.bitfire.synctools.storage.BatchOperation
 import at.bitfire.synctools.util.Utils.capitalize
-import at.bitfire.synctools.util.Utils.trimToNull
+import at.bitfire.synctools.util.trimToNull
 import at.bitfire.synctools.vcard.property.CustomType
 import ezvcard.parameter.RelatedType
 import java.util.LinkedList
