+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/contacts/builder/StructuredPostalBuilder.kt
@@ -8,7 +8,7 @@ import android.net.Uri
 import android.provider.ContactsContract.CommonDataKinds.StructuredPostal
 import at.bitfire.synctools.mapping.contacts.Contact
 import at.bitfire.synctools.storage.BatchOperation
-import at.bitfire.synctools.util.Utils.trimToNull
+import at.bitfire.synctools.util.trimToNull
 import ezvcard.parameter.AddressType
 import java.util.LinkedList
