+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/contacts/handler/RelationHandler.kt
@@ -7,7 +7,7 @@ package at.bitfire.synctools.mapping.contacts.handler
 import android.content.ContentValues
 import android.provider.ContactsContract.CommonDataKinds.Relation
 import at.bitfire.synctools.mapping.contacts.Contact
-import at.bitfire.synctools.util.Utils.trimToNull
+import at.bitfire.synctools.util.trimToNull
 import at.bitfire.synctools.vcard.property.CustomType
 import ezvcard.parameter.RelatedType
 import ezvcard.property.Related
