+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/tasks/builder/LocationBuilder.kt
@@ -6,7 +6,7 @@ package at.bitfire.synctools.mapping.tasks.builder
 
 import android.content.Entity
 import at.bitfire.ical4android.Task
-import at.bitfire.synctools.util.Utils.trimToNull
+import at.bitfire.synctools.util.trimToNull
 import org.dmfs.tasks.contract.TaskContract.Tasks
 
 class LocationBuilder : DmfsTaskFieldBuilder {
