+++ synctools/src/main/kotlin/at/bitfire/synctools/util/Utils.kt
@@ -18,12 +18,6 @@ object Utils {
         }
     }
 
-    /**
-     * Returns a string having leading and trailing whitespace removed.
-     * If the resulting string is empty, returns null.
-     */
-    fun String?.trimToNull(): String? = this?.trim()?.takeIf { it.isNotBlank() }
-
     fun StructuredName.isEmpty() =
         prefixes.isEmpty() && given == null && additionalNames.isEmpty() && family == null && suffixes.isEmpty()
